Refactor - removing duplicate interfaces
